replaced bioformats plugin with vcell-bioformats thrift service
Bioformats ImageDataset Service is an executable Jar which is downloaded
by user request.  Both vcell and vcell-bioformats projects share
identical .thrift files - so identical language bindings.  Service is
not linked into vcell, but is launched as a shell command and listens on
a socket for commands via thrift serialization and rpc.