Isolate the Oracle-specific code to its own module
With these changes, vanilla systems can now build all of VirtualCell
except for the new vcell-oracle component!

It was easy thanks to the pluginification of the databases. :-)

If I am reading the Oracle license agreement correctly, redistribution
of unmodified ojdbc6 and ucp JAR files seems allowed. We can discuss the
best way to proceed with that in a Maven-friendly way. One solution
could be to add them to the SciJava Maven repository, as long as doing
so is legal. Otherwise, we could use the non-maven-jar-maven-plugin:
  https://github.com/stephenc/non-maven-jar-maven-plugin