Improve ojdbc6 and ucp dependency handling
We now use the non-maven-jar-maven-plugin to manage them in a more
"Maven-like" way. It still causes trouble if someone tries to depend on
vcell-oracle transitively outside the multi-module build, but at least
you don't have to install these artifacts locally by hand anymore.

Note that the two new components ojdbc6 and ucp are not intended to be
opened in Eclipse. I tried adding an Eclipse-specific profile to fail
the build via the Enforcer if these components are open, but
unfortunately it does not report the details of how the enforcer failed.
So for now, you just have to know to avoid importing these projects into
Eclipse, or keep them closed.