Showing some helpful text in the biomodel Object Properties Panel.
Because the real properties panel for the biomodel is shown in the upper
right corner of the screen, and the annotations now have their own tab,
there isn't much that we can show here (lower right corner).
So one solution would be to show some help, or tip of the day...
Proof of concept code.