Animation's have been properly integrated. Bug related to generating  SpringSalad input file has been fixed.
I have made it so that the animation button and the animation itself only appear if run_animation() is called in PlotPane.Java. Right now it is being called when the ODE Viewer is being made if langevin solver is detected. This makes it so that it only runs while the results are going to be shown and only for SpringSaLaD applications.

I also changed the internal code of the ViewerGui.java file from SpringSaLaD so that it looks for the right folder because they are named differently than usual and there are multiple in the temp folder.

I found a bug where dt_data was written before dt_spring when creating the SpringSaLaD input file. I found that it was because dt_data was being written through the writeData() function in SolverTaskDescription whereas dt_spring was written after using the writeTimeInformation() method. To correct this, I moved the writing of dt_data to the writeTimeInformation() method so that I could order them appropriately and because it should also be classified as time information. I made this change to the writeTimeInformation() methods in SpringSaLaDExporter.java and LangevinLngvWriter.java.