Modifying S3Proxy Environmental Variables
Made it so that there is no CORS requests available, allowing only known headers, and require no authorization. There shouldn't be any other domains trying to access the resources in the S3 buckets since we'll me accessing it directly. If we want authorization then we need to make an AWS account which is responsible for this. Unknown headers should not be used since we'll be using already made API tools (AWS libraries/CLI) and they shouldn't act funny.