Made Exporter Integrate with Rest of VCell
The exporter previously only had it's API and export service function work with hard coded variables for simulation data paths. Now all of the hard coded variables utilize environmental variables that are set within the VCell containers. Also added S3 Proxy to the docker compose file so that it can now be utilized.