Transition from Dataset Controller to Data Server Impl
The previous method for retrieving sim data was to use the data set controller class, but that only allowed for user generated simulation results to be accessed. With data server impl it allows for all simulation results to be accessible as long as the user has permission to do so.