Support Output Functions
Output functions don't have a data identifier, so in the initial section of the code when I attempt to determine variable type a null pointer execution get's raised. Now that check also determines if the data identifier retrieved is a null pointer and if not the data for output functions is retrieved.