Proper Naming N5 Export Popup
Before the N5 export would use the word dataset and not give immediate feedback for improper N5 file naming. Now it does give proper feedback and it states that the user is naming the N5 file, when in reality it's naming a dataset within an N5 file, but they don't need to confuse themselves with that.