Write Custom Metadata
Before the additional MetaData that wanted to be written with the attributes never worked if it was a Map, so now I created a record that helps GSON to understand what needs to be written and can be expanded upon. Most of the entries follow what N5 metadata should be for ImageJ, and there are tests against N5 internal libraries to ensure nothing strange happens.