Proper Sequence of Events for VCell Client
There was a bug where the third task in the list of async tasks never got executed. This actually turned out to be a beneficial bug, because otherwise it would have resulted in a connection being created twice for the client. Now the third task does execute, and properly does the job it's supposed to do which is create connections.