HTC Memory Request Bug Fixed
Previously the amount of memory that could be requested was not bounded, for whatever the user requested was always allowed. Later in the process the requests would be rejected by slurm, but they still where made. Now they now longer are made if the request is above the maximum set.