Expanded Tracer for multi-threaded operation
Did that message concern you? Good, read this:

Basically, in order to have a proper timeout in CLI, we need a "watchdog" thread to time while another thread does the actual work. If the work takes too long, we choose to have the watch dog thread call an interrupt on the running thread, and the watchdog continues as the "main" thread.

The problem is that `Tracer` was intentionally desiged to keep data thread-independent. This is great, because then we know which thread did what work, and we don't need to worry about multi-threaded access.

But that means (before this commit), the watchdog would get 0 knowledge from the worker thread's Tracer. Meaning it wouldn't know the details of any failures (or even if the run failed! we don't check the return code, we check the **Tracer*)!

To remedy this, Tracer has been redesigned as type of "singleton", that when given a thread id, can fetch the tracer belonging to another thread.

!!!!!!!!!!!!!BUT!!!!!!!!!!!!! All methods that modify the Tracer in any way do so ONLY ON THE CURRENT THREAD'S INSTANCE.

We can NOT allow cross thread modification. Doing so opens a whole host of race conditions, complexity, etc. the Tracer was never supposed to deal with.

Comments have been left in the code to support this effort.